chore: rename file and update docs
